Discord/requirements.txt
 discord.py
 pip
 asyncio
 datetime
 python-dotenv
 json